New `tantan` module for detecting tandem repeats
The `tantan` software takes biological sequences in FASTA format as
input and detects low-complexity and tandem repeat sequences in them.

Upstream repository: https://gitlab.com/mcfrith/tantan

Primary publication: https://doi.org/10.1093/nar/gkq1212

`tantan` has 5 different types of output:

 * `-f0` for Masked sequences (FASTA format)
 * `-f1` for repeat probabilities (text format)
 * `-f2` for repeat counts (text format)
 * `-f3` for repeat coordinates (BED format)
 * `-f4` for tandem repeats (text format)

This module requires that its input sequence is not compressed.

By default, the module will output masked sequences in its `fasta`
output channel.  It also has a `BED` and a `txt` output channel for the
other formats.  The formats and channel are chosen by parsing the
command-line arguments for the `-f` option.  For this reason, the use of
a space between `-f` and the output type is not accepted and the module
will exit with error if it happens.